UX: removed displaying lose/win in gameover page after each game.
In game over page, comparing the current score with the high score, a text would be rendered saying 'You win' or 'You lose' which seem to be discouraging to a kid when he/she is just few scores behind the high score. Instead, just the score and high score would be displayed.